Feat: 카테고리 삭제 api에서 삭제하려는 카테고리에 상품이 있을 떄의 예외처리 추가

+++ src/main/java/com/example/shoppingmall_comp/domain/items/service/implement/CategoryServiceImpl.java
@@ -51,8 +51,13 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional
     public void delete(Long categoryId) {
-        categoryRepository.findById(categoryId)
+        Category category = categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND_CATEGORY));
+
+        if (itemRepository.findFirstByCategory(category).isPresent()) {
+            throw new BusinessException(ErrorCode.CATEGORY_EXIST_GOODS);
+        }
+
         categoryRepository.deleteById(categoryId);
     }
